[#25] refactor: EssayService에서 AI 관련 로직을 EssayAiService로 분리
- generate/regenerate/recommend를 EssayAiService로 이동
- EssayService는 CRUD 책임만 보유 (140줄로 축소)
- 외부 클라이언트(EmbeddingClient, LlmClient, EssayPromptBuilder) 의존성을 AI 서비스로 격리
- EssayController는 두 서비스를 모두 주입받아 라우팅
